+++ documentation/changelog.rst
@@ -1,10 +1,11 @@
 Changelog
 =========
 
-Juni XX, 20202 - bake-toolkit 2.63.0
+Juni 10, 2020 - bake-toolkit 2.63.0
     * Added: Internal developer feature "-Z dep-overview=<json>" for intelligent CIs.
-    * Added: Forbid include dirs with a leading space, e.g. *IncludeDir " include"* which is now treated as an error.
-    * Added: Files have an optional attribute "linkDirectly", which means they are not included in the lib, but linked directly to the executable.
+    * Added: Forbid include dirs with a leading space, e.g. *IncludeDir " include"*.
+    * Added: An info is printed out if path magic is used in IncludeDir.
+    * Added: Files have an optional attribute "linkDirectly", which means they are not archived in a lib, but linked directly to the executable.
 
 May 14, 2020 - bake-toolkit 2.62.0
      * Added: bake now understands dependency files from Axivion, which makes incremental build possible.